[#3280] 서비스 계층 구조 api에 accoun id 권한 검사 로직 적용

+++ BillingApiTres/Controllers/ServiceHierachies/DeleteServiceHierarchyController.cs
@@ -1,6 +1,8 @@
 ﻿using Billing.Data.Interfaces;
+using BillingApiTres.Models.Validations;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Collections.Immutable;
 
 namespace BillingApiTres.Controllers.ServiceHierachies
 {
@@ -8,9 +10,9 @@ namespace BillingApiTres.Controllers.ServiceHierachies
     [Authorize]
     public class DeleteServiceHierarchyController(
         IServiceHierarchyRepository serviceHierarchyRepository,
+        IConfiguration config,
         ILogger<DeleteServiceHierarchyController> logger) : ControllerBase
     {
-
         [HttpDelete("/service-organizations/{serialNo}")]
         public async Task<ActionResult> Delete(long serialNo)
         {
@@ -18,6 +20,10 @@ namespace BillingApiTres.Controllers.ServiceHierachies
 
             if (entity == null)
                 return NotFound(new { serialNo = serialNo });
+
+            var accountIds = HttpContext.Items[config["AccountHeader"]!] as ImmutableHashSet<long>;
+            if (accountIds?.Contains(entity.AccountId) == false)
+                return Forbid();
 
             await serviceHierarchyRepository.Delete(entity);
